catalog-service/requirements.txt
pymongo
Flask
python-dotenv
flask-cors
<<<<<<< HEAD
PyJWT
=======
PyJWT
requests
>>>>>>> b6eb336 (test)
